latest Update

+++ Agazaty/Models/AuthModel.cs
@@ -7,10 +7,10 @@ namespace Agazaty.Models
         public string FullName { get; set; }
         public string Email { get; set; }
         public string DepartmentName { get; set; }
-        public double SickLeavesCount { get; set; }
-        public double NormalLeavesCount { get; set; }
-        public double CasualLeavesCount { get; set; }
-        public double PermitLeavesCount { get; set; }
+        public int  SickLeavesCount { get; set; }
+        public int NormalLeavesCount { get; set; }
+        public int CasualLeavesCount { get; set; }
+        public int PermitLeavesCount { get; set; }
         public List<string> Roles { get; set; }
         public string Message { get; set; }
         public bool IsAuthenticated { get; set; }
